Move storage client logic into core.py
